Fix segfault on Alpine
A segfault was happening on Alpine build but not on other distributions.
It turns out it was due to a logical bug in the code.
`current_size` was updated regardless of the size of `length` compared
to it, instead of only being updated when `length` is superior to
`current_size`.